% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_get_data.R
\name{ach_get_data}
\alias{ach_get_data}
\title{ach_get_data}
\usage{
ach_get_data(date = NULL, dir = NULL)
}
\arguments{
\item{date}{folder name which corresponds to a date. If null, will pull most recent date}

\item{dir}{file path relative to working directory}
}
\description{
This function pulls ACHQC data.
}
\examples{
ach_get_data(date = "2021-11-08", dir = "../../../data")
}
