% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_to_df.R
\name{anova_to_df}
\alias{anova_to_df}
\title{anova_to_df}
\usage{
anova_to_df(anv, incl_total = F, p_nice = F)
}
\arguments{
\item{anv}{an anova.rms object}

\item{incl_total}{boolean to include total row in anova.rms}

\item{p_nice}{boolean. whether to print p-values nicely with "<0.001" and rounded numbers.}
}
\description{
Converts rms::anova object to data.frame
}
