% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_inc.R
\name{break_inc}
\alias{break_inc}
\title{Helper function for \code{ggplot}'s \code{scale_x/y_continuous}}
\usage{
break_inc(inc = 10, sf = 0.05, ...)
}
\arguments{
\item{inc}{increments}

\item{sf}{scaling factor (multiplicative expand) - higher \code{sf} will result in wider margins before ticks begin}
}
\description{
Allows faceted free_x and free_y scales to have same break increments
}
\examples{
library(ggplot2)
ggplot(data = iris, aes(x = Petal.Width, y = Sepal.Length)) +
  geom_point() +
  facet_grid( ~ Species, scales = "free") +
  scale_x_continuous(breaks=break_inc(inc=0.1, sf=.05),
                     expand = c(0.05, 0))
}
