% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_plot.R
\name{butterfly_plot}
\alias{butterfly_plot}
\title{butterfly_plot}
\usage{
butterfly_plot(
  x,
  y,
  col1 = "#FF000060",
  col2 = "#3200D360",
  border1 = "black",
  border2 = "black",
  breaks = 100,
  label = T,
  legend = F,
  legendloc = "bottomleft",
  flipaxis = F,
  psrange = NULL
)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired}

\item{y}{a vector of values for which groups are defined}

\item{col1}{character, color of histogram for group 1, default is #FF000060}

\item{col2}{character, color of histogram for group 2, default is #3200D360}

\item{border1}{character, color of histogram border of group 1, default is black}

\item{border2}{character, color of histogram border of group 2, default is black}

\item{breaks}{a single number giving the number of cells for the histogram, default is 100}

\item{label}{logical, whether to add group labels above histograms, default is TRUE}

\item{legend}{logical, whether to add legend, default is FALSE}

\item{legendloc}{location location specified by setting to a single keyword from the list "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". Default is "bottomleft."}

\item{flipaxis}{logical, direction of the butterfly. FALSE is butterfly shape i.e. reflected over y-axis, TRUE is reflected over x-axis. Default is TRUE.}

\item{psrange}{range of histogram values}
}
\description{
This function creates a butterfly plot commonly used to show propensity score overlap
}
\examples{
butterfly(x = data$ps, data$group, label = F, flipaxis = T)
butterfly(x = data$ps, data$group, label = F, flipaxis = F, breaks = 10)
}
