% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLists.R
\name{compareLists}
\alias{compareLists}
\title{Compile missing data summary of data set}
\usage{
compareLists(list1, list2, lnames)
}
\arguments{
\item{list1}{a list}

\item{list2}{a list}

\item{lnames}{a vector of length 2 supplying names of list}
}
\description{
Compares lists element-wise. Created for purpose of comparing two lists consisting of data frames, where each list is from a different data-pull date.
}
\details{
The output is a text summary of element-wise list comparisons
}
\examples{
l1 <- list(tmp1 = data.frame(x = 1:100, y = 100:1), tmp2 = data.frame(a = letters, b = LETTERS))
l2 <- list(tmp1 = data.frame(x = 1:99, y = 99:1), tmp2 = data.frame(a = letters[1:14], b = LETTERS[1:14]))
compareLists(l1, l2)
compareLists(l2, l1)
}
