% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_breaks.R
\name{equal_breaks}
\alias{equal_breaks}
\title{Helper function for \code{ggplot}'s \code{scale_x/y_continuous}}
\usage{
equal_breaks(n = 3, s = 0.05, ...)
}
\arguments{
\item{n}{number of breaks}

\item{sf}{scaling factor (multiplicative expand) - higher \code{sf} will result in wider margins before ticks begin}
}
\description{
Allows faceted free_x and free_y scales to have equal number of breaks with differing ranges
}
\examples{
library(ggplot2)
ggplot(data = iris, aes(x = Petal.Width, y = Sepal.Length)) +
  geom_point() +
  facet_grid( ~ Species, scales = "free") +
  scale_x_continuous(breaks=equal_breaks(n=3, sf=.05),
                     expand = c(0.05, 0))
}
