% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvn.R
\name{fvn}
\alias{fvn}
\title{fvn}
\arguments{
\item{pattern}{A character string regex search expression}

\item{data}{A data.frame object}
}
\description{
Short for "find variable name." List variable names / vector items which match a search expression
}
\keyword{fvn}
