% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fmt.R
\name{get_fmt}
\alias{get_fmt}
\title{Prints out preferred HTML formatting for RMarkdown documents}
\usage{
get_fmt(os = "PC")
}
\arguments{
\item{os}{Operating system - either "PC" or "mac." Will adjust CSS according to operating system.}
}
\description{
Prints out preferred HTML formatting for RMarkdown documents. Will print YAML, CSS styling, global chunk options, and common packages. Currently not customizable via the function. Purpose is to print out
text to copy/paste into new RMarkdown documents for quick access to favorite style.
This is a preferred alternative to Rmarkdown templates as customizable features may be included in the future.
}
