% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_two_column_consort.R
\name{gg_two_column_consort}
\alias{gg_two_column_consort}
\title{gg_two_column_consort}
\usage{
gg_two_column_consort(data, var, top_box_text, date = NULL)
}
\arguments{
\item{data}{data.frame, data with consort variable}

\item{var}{string, name of consort variable that is in the format "col row description"}

\item{top_box_text}{string, text to display in top box of diagram}

\item{date}{string, date of data pull. if NULL, won't display date.}

\item{box_height}{height of exclusion boxes. default is 0.4.}

\item{box_width}{width of subgroup boxes. default is 0.22.}
}
\description{
This function creates a two column consort figure with "before" and "after" population
N If you would also like to display subgroup population Ns, please use
\code{gg_two_column_consort_subgroup}
}
