% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_two_column_consort_subgroup.R
\name{gg_two_column_consort_subgroup}
\alias{gg_two_column_consort_subgroup}
\title{gg_two_column_consort_subgroup}
\usage{
gg_two_column_consort_subgroup(
  data,
  var,
  var2,
  box_width = 0.4,
  box_height = 0.4,
  box2_width = 0.22,
  box2_height = 0.8,
  top_box_text,
  date = NULL
)
}
\arguments{
\item{data}{data.frame, data with consort variable}

\item{var}{string, name of consort variable that is in the format "col row description"}

\item{var2}{string, name of consort variable for subgroup population that is in the format "col row description"}

\item{box_width}{width of subgroup boxes. default is 0.22.}

\item{box_height}{height of subgroup boxes. default is 0.8.}

\item{top_box_text}{string, text to display in top box of diagram}

\item{date}{string, date of data pull. if NULL, won't display date.}
}
\description{
This function creates a two column consort figure with "before" and "after" population
N as well as subgroup population Ns.
}
