% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot.R
\name{kmplot}
\alias{kmplot}
\title{Create a Kaplan Meier plot}
\usage{
kmplot(
  df,
  km,
  colors,
  conf.int = T,
  p.val = F,
  break.int = 6,
  risk.table = "nrisk_cumcensor",
  cumevents = T,
  lgd = "top",
  ttl = "",
  ylab = "",
  xlab = "",
  height = 10,
  surv.median.line = "hv",
  labs,
  text.size = 14,
  title.size = 16,
  legend.size = 12
)
}
\arguments{
\item{df}{a data set used to fit survival curves}

\item{km}{a survfit object}

\item{colors}{a vector of colors for survival curves, length equal to number of curves.}

\item{conf.int}{logical. if TRUE, plots confidence interval. Default is TRUE.}

\item{p.val}{logical. if TRUE, adds p-value to the plot. Default is FALSE.}

\item{break.int}{numeric. controls when time axis ticks are presented. Defaults to 6 units (often used for 6 months).}

\item{risk.table}{logical, or string. allowed values are those allowed in the \code{risk.table} in \code{survminer::ggsurvplot}. Default is "nrisk_cumcensor".}

\item{cumevents}{logical. logical value specifying whether to show or not the table of the cumulative number of events. Default is TRUE.}

\item{lgd}{character specifying the legend location. Default is "top"}

\item{ttl}{legend title}

\item{ylab}{y-axis label}

\item{xlab}{x-axis label}

\item{height}{height of the survival plot. Default is 10. Ignored when risk.table = FALSE}

\item{surv.median.line}{a character for how median survival lines should be drawn. Allowed values include one of c("none", "hv", "h", "v"). v: vertical, h:horizontal.}

\item{labs}{legend labels. If not specified, will use strata names.}

\item{text.size}{text size for axis text}

\item{title.size}{text size for plot titles}

\item{legend.size}{text size for legend}
}
\description{
Creates a Kaplan Meier plot with the option of risk and cumulative event tables. This is essentially my favorite KM plot using the \code{survminer} package. Default is my favorite settings.
}
\examples{
kmplot(df = lung, km =survfit(Surv(time, status) ~ sex, data = lung), surv.median.line = "hv", break.int = 365)

survObj <- survfit(Surv(time, status) ~ ph.ecog, data = lung)
kmplot(df = lung, km = survObj, surv.median.line = "hv", break.int = 180, colors = c("red", "blue", "orange", "black"), labs = gsub("ph.ecog=", "", attr(survObj$strata, "names")))
}
\keyword{ggplot}
\keyword{kaplan}
\keyword{meier,}
