% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit2prob.R
\name{logit2prob}
\alias{logit2prob}
\title{logit2prob}
\usage{
logit2prob(logit)
}
\arguments{
\item{logit}{a vector of values on logit scale}
}
\description{
This function converts a value on logit scale into a probability
}
\examples{
logit2prob(logit = data$logit)
}
