% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_summary.R
\name{missing_summary}
\alias{missing_summary}
\title{Compile missing data summary of data set}
\usage{
missing_summary(data, upper_limit, max_vars, include_vars, type = "both")
}
\arguments{
\item{data}{A data.frame object for which a missing data report will be generated}

\item{upper_limit}{A number.  The right tail of the frequency distribution reported in the \code{type="row"} summary is truncated to "upper_limit +".}

\item{max_vars}{A number.  Limits the list of variables reported \code{type="col"} to the first \code{max_vars} most frequently missing variables.  If there are multiple variables with the same number of missing values, all such the variables will be reported. (This means more than max_vars variables can appear in the output)}

\item{include_vars}{A vector of variable names or a regular expression to select variables that match a pattern. You may drop variables by providing a regular expression preceded by \code{!} (\code{include_vars = "!qol"}, for example, would drop variables matched by "qol")}

\item{type}{Type of output. One of \code{"both"}, \code{"row"}, \code{"col"}, or \code{"complete"}. The default is \code{"both"} for both \code{"row"}-wise and \code{"col"}-wise missing data summaries. \code{type="complete"} will report proportion of records with complete data.}
}
\description{
Generates missing data summaries.  Adapted from \code{tgsify}'s \code{missingness_info}
}
\details{
The output is a list with the "by row" summary and the "by column" summary.
}
\examples{
missing_summary(airquality)
missing_summary(airquality, type = "complete")
missing_summary(airquality, type = "row")
missing_summary(airquality, type = "col")
Include only Ozone and Solar.R variables
missing_summary(airquality, type = "row", include_vars = c("Ozone", "Solar.R"))
missing_summary(airquality, type = "row", include_vars = "Oz|Solar")
missing_summary(airquality, type = "both", include_vars = "Oz|Solar")
missing_summary(airquality, type = "row", upper_limit = 1)
Below, the upper_limit will provide the same results
missing_summary(airquality, type = "row", upper_limit = 3)
upper_limit = 6 will not return "N+" like the above example because `airquality` has 6 total variables
missing_summary(airquality, type = "row", upper_limit = 6)
drop Ozone only
missing_summary(airquality, type = "col", include_vars = "!Ozone")
}
\keyword{missingness_info}
