% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msheets.R
\name{read_msheets}
\alias{read_msheets}
\title{Read in Excel workbook with multiple sheets without using Java}
\usage{
read_msheets(filename)
}
\arguments{
\item{filename}{increments}

\item{sf}{scaling factor (multiplicative expand) - higher \code{sf} will result in wider margins before ticks begin}
}
\value{
data.frame
}
\description{
Read in Excel workbook with multiple sheets without using Java. Uses \code{readxl}
}
\examples{
\dontrun{
read_msheets(here("myexcel.xlsx"))
}
}
