% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMac.R
\name{renderMac}
\alias{renderMac}
\title{renderMac}
\usage{
renderMac(filename)
}
\arguments{
\item{filename}{Character string that is the file name.}
}
\description{
Renders file on mounted drive when running on a Mac. Must set working directory to file location.
}
\details{
Rendered file will be put in "~/Desktop"

Must set working directory to file location.
}
\examples{
# Not run:
}
\keyword{render}
