% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubin.R
\name{rubin}
\alias{rubin}
\title{rubin}
\usage{
rubin(df)
}
\arguments{
\item{df}{a data.frame of minimum 3 columns with the following information in order: variable name, coefficient, std. error}
}
\description{
Creates regression model statistics for multiple imputation using Rubin's rules.
}
