% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangram_to_df.R
\name{tangram_to_df}
\alias{tangram_to_df}
\title{tangram_to_df}
\usage{
tangram_to_df(tbl, add_rown = FALSE, html = TRUE)
}
\arguments{
\item{tbl}{a \code{tangram} object (table)}

\item{add_rown}{logical. Whether to add a row id (\code{id}) to the output dataframe}

\item{html}{logical. Whether to include html code for indentation.}
}
\description{
This function converts a \code{tangram} object to dataframe. Will be in html format.
}
\examples{
library(tangram)
tab <- tangram(drug~bili+albumin+stage::Categorical+protime+sex+age+spiders,
              data = pbc,
              style = "lancet",
              id = "mytbl3")
tangram_to_df(tbl = tab, html = F)
}
